add a meter for the distance ride by the player

+++ MaBibliotheque/Draw.hpp
@@ -8,7 +8,7 @@ public:
 	SFML_output();
 	void display(Route& route);
 	void clean();
-	void affichage_text(sf::Font& font, sf::String string, int size_char, bool bold, bool underlined, double pos_x, double pos_y);
+	void affichage_text(sf::Font& font, sf::String string, int size_char, bool bold, bool underlined, float pos_x, float pos_y);
 	void affichage_game_over(Route& route, sf::Font& font);
 	void affichage_donnees_joueur_en_jeu(Route& route, sf::Font& font);
